Fix library search path for man page generation
This patch will fix an issue that occurs when generating the man pages
for swift-im. The man page is generated from the Swiften library using
help2man. The issue occurs when the library cannot be found during build
time and help2man is generating a man page with the output of that
error, ignoring the return error code from swift-im. To avoid this
happening in future builds, swift-im --help is run before generating the
page and checked if the execution was successful. If yes we proceed with
the generation, if not we stop the build and throw an error.
swiften-config man page generation was not affected by this issue.

Test-Information:

Build with ubuntu 18.04 on a machine that hasn’t got Swift or Swiften
installed. Build the Debian package for Debian stretch with pbuilder,
and installed on a clean docker container (Debian stretch). Repeated the
test with Debian sid but with packages build from docker containers.

Change-Id: Icf8a84ba7f4810c87e5835e5ac1a91d1324ddac8